requirements.txt
numpy
scipy
opencv-python
discord.py
python-dotenv
Pillow==10.2.0
imagehash==4.3.1
